added test

+++ screwbox-core/src/main/java/dev/screwbox/core/environment/softphysics/RopeRenderComponent.java
@@ -18,7 +18,7 @@ public class RopeRenderComponent implements Component {
     /**
      * Stroke width used for drawing.
      */
-    public int strokeWidth;
+    public double strokeWidth;
 
     /**
      * Color used for drawing.
@@ -30,7 +30,7 @@ public class RopeRenderComponent implements Component {
      */
     public int drawOrder;
 
-    public RopeRenderComponent(final Color color, final int strokeWidth) {
+    public RopeRenderComponent(final Color color, final double strokeWidth) {
         this.color = color;
         this.strokeWidth = strokeWidth;
     }
